smarter sort finished, 20 moves less on same test input, rapid sort on the way

+++ src/stacks_functions2.c
@@ -6,14 +6,14 @@
 /*   By: rcochran <rcochran@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/14 14:05:58 by rcochran          #+#    #+#             */
-/*   Updated: 2025/02/18 19:39:20 by rcochran         ###   ########.fr       */
+/*   Updated: 2025/02/25 11:13:11 by rcochran         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "push_swap.h"
 
-int		ft_stacksize(t_stack *stack);
-t_stack	*ft_stacklast(t_stack **stack);
+long		ft_stacksize(t_stack *stack);
+t_stack		*ft_stacklast(t_stack **stack);
 
 /* get stack last item */
 t_stack	*ft_stacklast(t_stack **stack)
@@ -31,9 +31,9 @@ t_stack	*ft_stacklast(t_stack **stack)
 }
 
 /* get stack size */
-int	ft_stacksize(t_stack *stack)
+long	ft_stacksize(t_stack *stack)
 {
-	int	size;
+	long	size;
 
 	size = 0;
 	while (stack != NULL)
